IT-3686: Dedup tag keys
