+++ set_tags/set_instance_tags.py
@@ -83,7 +83,12 @@ def create_or_update(event, context):
   extra_tags.append(provisioned_product_name_tag)
   access_approved_role_tag = utils.get_access_approved_role_tag(instance_tags)
   extra_tags.append(access_approved_role_tag)
-  all_tags = list(synapse_tags + extra_tags)
+  # 
+  # In case of duplication, Synapse tags should replace existing
+  # bucket tags.
+  #
+  all_tags = utils.merge_tags(extra_tags, synapse_tags)
+  
 
   volume_ids = get_volume_ids(instance_id)
   log.debug(f'Apply tags: {all_tags} to volume {volume_ids}')
